Extract file uploading logic
