+++ front/app/containers/Admin/projects/new/ProjectSetupForm.tsx
@@ -5,8 +5,6 @@ import { isEmpty } from 'lodash-es';
 import { Multiloc, UploadFile, CLErrors } from 'typings';
 
 import { IFileAttachmentData } from 'api/file_attachments/types';
-import { IFileData } from 'api/files/types';
-import useAddFile from 'api/files/useAddFile';
 import useAuthUser from 'api/me/useAuthUser';
 import useAddProjectImage from 'api/project_images/useAddProjectImage';
 import useDeleteProjectImage from 'api/project_images/useDeleteProjectImage';
@@ -54,17 +52,16 @@ import {
 import SpaceSelectSection from 'components/admin/SpaceSelectSection';
 import SubmitWrapper, { ISubmitState } from 'components/admin/SubmitWrapper';
 import Highlighter from 'components/Highlighter';
-import FileRepositorySelectAndUpload from 'components/UI/FileRepositorySelectAndUpload';
 import Warning from 'components/UI/Warning';
 
 import { FormattedMessage, useIntl } from 'utils/cl-intl';
 import { queryClient } from 'utils/cl-react-query/queryClient';
 import clHistory from 'utils/cl-router/history';
 import Link from 'utils/cl-router/Link';
-import { generateTemporaryFileAttachment } from 'utils/fileUtils';
 import { isNilOrError } from 'utils/helperUtils';
 import { isSpaceModerator } from 'utils/permissions/roles';
 
+import FileUploader from '../_shared/components/ProjectSetupForm/FileUploader';
 import { TOnProjectAttributesDiffChangeFunction } from '../_shared/types';
 
 const FOLDER_SELECT_ALLOWED_HIGHEST_ROLES: (string | undefined)[] = [
@@ -91,7 +88,6 @@ const ProjectSetupForm = () => {
   const syncProjectFiles = useSyncFiles();
 
   // File Attachments
-  const { mutate: addFile, isLoading: isAddingFile } = useAddFile();
   const [projectFileAttachments, setProjectFileAttachments] = useState<
     IFileAttachmentData[] | undefined
   >();
@@ -180,82 +176,6 @@ const ProjectSetupForm = () => {
   const handleProjectCardImageOnCompleteCropping = (base64: string) => {
     setSubmitState('enabled');
     setCroppedProjectCardBase64(base64);
-  };
-
-  const handleProjectFileOnAdd = (fileToAdd: UploadFile) => {
-    // Upload the file to the Data Repository, so we can make the attachment later.
-    addFile(
-      {
-        content: fileToAdd.base64,
-        name: fileToAdd.name,
-        category: 'other', // Default to 'other' when added from phase setup
-        ai_processing_allowed: false, // Default to false when added from phase setup
-      },
-      {
-        onSuccess: (newFile) => {
-          // Create a temporary file attachment to add to the state, so the user sees it in the list.
-          const temporaryFileAttachment = generateTemporaryFileAttachment({
-            fileId: newFile.data.id,
-            attachableId: undefined,
-            attachableType: 'Project',
-            position: projectFileAttachments
-              ? projectFileAttachments.length
-              : 0,
-          });
-
-          const isDuplicate = projectFileAttachments?.some((fileAttachment) => {
-            return (
-              fileAttachment.relationships.file.data.id ===
-              temporaryFileAttachment.relationships.file.data.id
-            );
-          });
-
-          setProjectFileAttachments(
-            isDuplicate
-              ? projectFileAttachments
-              : [...(projectFileAttachments || []), temporaryFileAttachment]
-          );
-
-          setSubmitState(isDuplicate ? submitState : 'enabled');
-        },
-      }
-    );
-  };
-
-  const handleProjectFileOnRemove = (
-    projectFileToRemove: IFileAttachmentData
-  ) => {
-    setProjectFileAttachments((projectFileAttachments) =>
-      projectFileAttachments?.filter(
-        (fileAttachment) => fileAttachment.id !== projectFileToRemove.id
-      )
-    );
-    setProjectFileAttachmentsToRemove((projectFileAttachmentsToRemove) => [
-      ...projectFileAttachmentsToRemove,
-      projectFileToRemove,
-    ]);
-
-    setSubmitState('enabled');
-  };
-
-  const handleFilesReorder = (updatedFiles: IFileAttachmentData[]) => {
-    setProjectFileAttachments(updatedFiles);
-    setSubmitState('enabled');
-  };
-
-  const handleProjectFileOnAttach = (fileToAttach: IFileData) => {
-    const temporaryFileAttachment = generateTemporaryFileAttachment({
-      fileId: fileToAttach.id,
-      attachableId: undefined,
-      attachableType: 'Project',
-      position: projectFileAttachments ? projectFileAttachments.length : 0,
-    });
-
-    setProjectFileAttachments((projectFileAttachments) => [
-      ...(projectFileAttachments || []),
-      temporaryFileAttachment,
-    ]);
-    setSubmitState('enabled');
   };
 
   const handleTopicsChange = (topicIds: string[]) => {
@@ -556,17 +476,17 @@ const ProjectSetupForm = () => {
                 }
               />
             </SubSectionTitle>
-            <FileRepositorySelectAndUpload
-              id="project-edit-form-file-uploader"
-              onFileAdd={handleProjectFileOnAdd}
-              onFileRemove={handleProjectFileOnRemove}
-              onFileReorder={handleFilesReorder}
-              onFileAttach={handleProjectFileOnAttach}
-              fileAttachments={projectFileAttachments}
-              enableDragAndDrop
+            <FileUploader
+              projectFileAttachments={projectFileAttachments}
+              setProjectFileAttachments={(...args) => {
+                setSubmitState('enabled');
+                setProjectFileAttachments(...args);
+              }}
+              setProjectFileAttachmentsToRemove={(...args) => {
+                setSubmitState('enabled');
+                setProjectFileAttachmentsToRemove(...args);
+              }}
               apiErrors={apiErrors}
-              maxSizeMb={50}
-              isUploadingFile={isAddingFile}
             />
           </StyledSectionField>
         </Section>
